requirements.txt
python-dotenv
langchain
langchain-community
langchain-core
openai
slack-sdk 
slack-bolt 
flask